less memory for A*

+++ Source/UsedTypes.h
@@ -1,6 +1,6 @@
 #ifndef USED_TYPES
 #define USED_TYPES
 
-typedef double FPType;
+typedef long double FPType;
 
 #endif
